backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
twilio
